refactored: create MethodSignature, decomposite some packages
