+++ src/main/java/ua/artcode/dao/SimpleTaskDao.java
@@ -2,7 +2,7 @@ package ua.artcode.dao;
 
 import ua.artcode.exception.AppException;
 import ua.artcode.exception.NoSuchTaskException;
-import ua.artcode.model.CodingBatTask;
+import ua.artcode.model.codingbat.CodingBatTask;
 
 import java.util.List;
