+++ src/main/java/ua/artcode/dao/SimpleUserDao.java
@@ -2,8 +2,7 @@ package ua.artcode.dao;
 
 import ua.artcode.exception.NoSuchUserException;
 import ua.artcode.exception.UserAccountExistException;
-import ua.artcode.exception.UserAuthenticationFailException;
-import ua.artcode.model.UserAccount;
+import ua.artcode.model.common.UserAccount;
 
 import java.util.Set;
