+++ src/main/java/ua/artcode/dao/SimpleUserDaoImpl.java
@@ -4,7 +4,7 @@ import ua.artcode.db.UserAccountContainer;
 import ua.artcode.exception.NoSuchUserException;
 import ua.artcode.exception.UserAccountExistException;
 import ua.artcode.exception.UserAuthenticationFailException;
-import ua.artcode.model.UserAccount;
+import ua.artcode.model.common.UserAccount;
 
 import java.util.Set;
