+++ src/main/java/ua/artcode/db/CodingBatTaskContainer.java
@@ -1,10 +1,9 @@
 package ua.artcode.db;
 
-import ua.artcode.model.CodingBatTask;
+import ua.artcode.model.codingbat.CodingBatTask;
 
 import java.io.Serializable;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
