--- src/main/java/ua/artcode/model/TaskTestData.java
@@ -1,9 +1,0 @@
-package ua.artcode.model;
-
-/**
- * Created by serhii on 06.11.15.
- */
-public class TaskTestData {
-
-
-}
