+++ src/main/java/ua/artcode/model/common/AccountType.java
@@ -1,4 +1,4 @@
-package ua.artcode.model;
+package ua.artcode.model.common;
 
 public enum AccountType {
         ADMINISTRATOR, USER, MASTER, STUDENT, TEACHER, JIDAI;
