+++ src/main/java/ua/artcode/utils/CodingBatTaskGrabber.java
@@ -6,7 +6,7 @@ import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
-import ua.artcode.model.CodingBatTask;
+import ua.artcode.model.codingbat.CodingBatTask;
 
 import java.io.IOException;
 import java.util.ArrayList;
