+++ src/test/java/ua/artcode/service/TaskServiceTest.java
@@ -1,7 +1,6 @@
 package ua.artcode.service;
 
 import org.apache.log4j.Logger;
-import org.junit.After;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -11,10 +10,9 @@ import ua.artcode.dao.SimpleTaskDaoImpl;
 import ua.artcode.db.CodingBatTaskContainer;
 import ua.artcode.exception.AppException;
 import ua.artcode.exception.NoSuchTaskException;
-import ua.artcode.model.CodingBatTask;
+import ua.artcode.model.codingbat.CodingBatTask;
 
 import java.util.Arrays;
-import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
