+++ src/test/java/ua/artcode/utils/TestTaskJsonSerializer.java
@@ -3,7 +3,7 @@ package ua.artcode.utils;
 import jdk.nashorn.internal.runtime.Source;
 import org.junit.*;
 import org.junit.runners.MethodSorters;
-import ua.artcode.model.CodingBatTask;
+import ua.artcode.model.codingbat.CodingBatTask;
 
 import java.io.File;
 import java.io.IOException;
